pyproject.toml
[tool.poetry]
name = "dozer"
version = "0.1.10"
description = "Python wrapper for the Dozer API"
authors = ["Dozer Team <api@getdozer.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.51.3"
protobuf = "^4.22.0"
pyarrow = "^11.0.0"
grpcio-reflection = "^1.51.3"
polars = "^0.16.10"
tqdm = "^4.65.0"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.51.3"
autopep8 = "^2.0.1"
pytest = "^7.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
